+++ modules/aws-ecs-app/certificates.tf
@@ -1,10 +1,15 @@
+/*
+ * This file contains the configuration for AWS ACM certificates and Route 53 records.
+ * It creates an ACM certificate for the application load balancer and a separate one for the CDN.
+ * The certificates are validated using DNS validation method.
+*/
 resource "aws_acm_certificate" "acm-certificate" {
   domain_name       = "alb.${var.service_name}.${var.environment}.${data.aws_route53_zone.provided-zone.name}"
   validation_method = "DNS"
 
   validation_option {
     domain_name       = "alb.${var.service_name}.${var.environment}.${data.aws_route53_zone.provided-zone.name}"
-    validation_domain = "${data.aws_route53_zone.provided-zone.name}"
+    validation_domain = data.aws_route53_zone.provided-zone.name
   }
 }
 
@@ -26,22 +31,22 @@ resource "aws_route53_record" "route53-record" {
 }
 
 resource "aws_acm_certificate" "cdn-acm-certificate" {
-  provider = "aws.us-east-1"
+  provider          = aws.us-east-1
   domain_name       = "${var.service_name}.${var.environment}.${data.aws_route53_zone.provided-zone.name}"
   validation_method = "DNS"
 
   validation_option {
     domain_name       = "${var.service_name}.${var.environment}.${data.aws_route53_zone.provided-zone.name}"
-    validation_domain = "${data.aws_route53_zone.provided-zone.name}"
+    validation_domain = data.aws_route53_zone.provided-zone.name
   }
 }
 
 resource "aws_route53_record" "cdn-route53-record" {
   for_each = {
     for dvo in aws_acm_certificate.cdn-acm-certificate.domain_validation_options : dvo.domain_name => {
-      name   = dvo.resource_record_name
-      record = dvo.resource_record_value
-      type   = dvo.resource_record_type
+      name     = dvo.resource_record_name
+      record   = dvo.resource_record_value
+      type     = dvo.resource_record_type
       provider = "aws.us-east-1"
     }
   }