+++ modules/aws-ecs-app/provider.tf
@@ -4,8 +4,9 @@ terraform {
 
   required_providers {
     aws = {
-      source  = "hashicorp/aws"
-      version = "~> 5.0"
+      source                = "hashicorp/aws"
+      version               = "~> 5.0"
+      configuration_aliases = [aws.us-east-1]
     }
   }
 }
@@ -16,5 +17,5 @@ provider "aws" {
 
 provider "aws" {
   region = "us-east-1"
-  alias =  "us-east-1"
+  alias  = "us-east-1"
 }