+++ CHANGELOG.md
@@ -10,6 +10,23 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 
+## [v0.1.0] - 2024-04-13
+
+### Added
+
+- Output the HTTPS Custom domain URL
+- Added module description in README.md
+- Added missing comments
+
+### Fixed
+
+### Changed
+
+- Forced HTTPS for both CDN distribution and ALB flows
+- Refactoring
+
+### Removed
+
 ## [v0.1-beta] - 2024-04-12
 
 ### Added
@@ -45,5 +62,6 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 
-[v0.1-alfa]: https://github.com/thecillu/devops-aws-tf/releases/tag/v0.1-alfa
+[v0.1.0]: https://github.com/thecillu/devops-aws-tf/compare/v0.1-alfa...v0.1.0
 [v0.1-beta]: https://github.com/thecillu/devops-aws-tf/compare/v0.1-alfa...v0.1-beta
+[v0.1-alfa]: https://github.com/thecillu/devops-aws-tf/releases/tag/v0.1-alfa